feat(frontend): move theme picker to a Settings page, add account menu
Replaces the toolbar's light/system/dark button-toggle group and
plain "Sign in"/"Sign out" text buttons with a single avatar icon
button opening a menu (Settings, Sign out/in) — a much lighter
toolbar. Theme state moves out of AppComponent into a ThemeService so
both the toolbar (which still needs to apply the dark/light body
class from app start) and the new /settings page can share it.
