+++ frontend/src/app/app.component.html
@@ -16,45 +16,31 @@
         <span>COST CONTROL</span>
     }
     <span class="spacer"></span>
-    <mat-button-toggle-group
-        class="theme-toggle mr-4"
-        [value]="themeMode()"
-        (change)="setThemeMode($event.value)"
-        hideSingleSelectionIndicator
+    <button
+        mat-icon-button
+        aria-label="Account menu"
+        matTooltip="Account"
+        [matMenuTriggerFor]="userMenu"
     >
-        <mat-button-toggle value="light" matTooltip="Light mode">
-            <mat-icon>light_mode</mat-icon>
-        </mat-button-toggle>
-        <mat-button-toggle value="system" matTooltip="Follow system setting">
-            <mat-icon>brightness_auto</mat-icon>
-        </mat-button-toggle>
-        <mat-button-toggle value="dark" matTooltip="Dark mode">
-            <mat-icon>dark_mode</mat-icon>
-        </mat-button-toggle>
-    </mat-button-toggle-group>
-    @if (isAuthenticated()) {
-        <button
-            mat-button
-            class="user-icon"
-            aria-label="Logout"
-            matTooltip="Sign out"
-            (click)="signOut()"
-        >
-            Sign out
-            <mat-icon>logout</mat-icon>
-        </button>
-    } @else {
-        <button
-            mat-button
-            class="user-icon"
-            aria-label="Login"
-            matTooltip="Sign in"
-            (click)="signIn()"
-        >
-            Sign in
-            <mat-icon>login</mat-icon>
-        </button>
-    }
+        <mat-icon>account_circle</mat-icon>
+    </button>
+    <mat-menu #userMenu="matMenu">
+        <a mat-menu-item routerLink="/settings">
+            <mat-icon>settings</mat-icon>
+            <span>Settings</span>
+        </a>
+        @if (isAuthenticated()) {
+            <button mat-menu-item (click)="signOut()">
+                <mat-icon>logout</mat-icon>
+                <span>Sign out</span>
+            </button>
+        } @else {
+            <button mat-menu-item (click)="signIn()">
+                <mat-icon>login</mat-icon>
+                <span>Sign in</span>
+            </button>
+        }
+    </mat-menu>
 </mat-toolbar>
 
 <mat-sidenav-container>
