+++ frontend/src/app/settings/settings.component.scss
@@ -1,0 +1,23 @@
+.settings-card {
+    max-width: 480px;
+}
+
+.card-title {
+    display: flex;
+    align-items: center;
+    gap: var(--space-2);
+    font: var(--mat-sys-title-small);
+
+    mat-icon {
+        font-size: 18px;
+        width: 18px;
+        height: 18px;
+        color: var(--mat-sys-primary);
+    }
+}
+
+.setting-label {
+    font: var(--mat-sys-body-medium);
+    color: var(--mat-sys-on-surface-variant);
+    margin: 0 0 var(--space-3);
+}
